Update README.md
Diabetes present or not?

Introduction

Diabetes is a disease in which there is abnormal high level of glucose in the body as a result of the body inability to respond to the natural insulin hormones in the body. This is a dataset originally gotten from the National Institute for Diabetes and Digestive and Kidney Disease. It comprises of female of 20 years and above. Several data were collected and the dataset have 9 columns with 768 rows/entries. The column comprises of data such as pregnancies; to express the number of pregnanacies, glucose; to express the glucose level in the blood, blood pressure; to express blood pressure measurement, skin thickness;to express the thickness of the skin, insulin;to express insulin level in blood, BMI; to express the body mass index, diabetespedigreefunction; to express diabetes percentage, age; to express the age, outcome; to express the final result, 1 means yes, while 0 means no.

Objectives

To diagnostically predict whether a patient has disbetes based on certain diagnostic measurement in the dataset.

Findings

The analysis show that the glucose level in the blood determines whether a patient has diabetes or not. Also, age, insulin, diabetes- pedigreefunction, and BMI have little influence on whether a patient will be diagnosed with diabetes or not, while the other parameters show no relationship with the diagnosis of diabetes in a patient. 

Conclusion

Diagnosis of diabetes based on various diagnostic measurements depicts that elderly patients are more likely to be diagnosed with diabetes than younger ones and obese patients are at risk of being diagnosed with diabetes than those whose BMI is within a healthy range.